formula change

+++ src/CreditPaymentService.java
@@ -1,8 +1,8 @@
 public class CreditPaymentService {
-    public float calculate(float amountOfCredit, float creditTerm, float creditRate) {
-        float loanRatePerMonth = (creditRate / 100) / 12;
-        float monthlyPayment = amountOfCredit * (loanRatePerMonth + (loanRatePerMonth /
-                (((float) Math.pow( 1 + loanRatePerMonth, creditTerm) - 1))));
+    public double calculate(double amountOfCredit, double creditTerm, double creditRate) {
+        double loanRatePerMonth = (creditRate / 100) / 12;
+        double monthlyPayment = amountOfCredit * (loanRatePerMonth + (loanRatePerMonth /
+                (Math.pow( 1 + loanRatePerMonth, creditTerm) - 1)));
 
         return monthlyPayment;
     }
